Add files via upload
The last exercise from chapter 4 of Programming Principles and Practice Using C++ by Bjarne Stroustrup.
Here the user is prompted to input a name and a score, both being saved in two different vectors. After the user inputs "noname 0" the program lists all the users and their scores and gives the user the option to search the list either by name or by score.